scraper/requirements.txt
playwright==1.58.0
Jinja2==3.1.6
python-dotenv==1.1.0
